chore: Move keyring types and initial state to keyring package
