+++ packages/sysweb3-utils/src/index.ts
@@ -2,13 +2,11 @@ export * from './contracts';
 export * from './encryptor';
 export * from './tokens';
 export * from './validation';
-export * from './initialState';
 export * from './fee';
 export * from './transactions';
 export * from './trezor';
 export * from './signer';
 export * from './accounts';
-export * from './keyring';
 export * from './networks';
 export * from './spsbt';
 export * from './signer-info';
